app/requirements.txt
# Web Framework and Server
fastapi
uvicorn[standard]

# Telegram Bot Library
aiogram

# Database
sqlmodel
asyncpg

# Reporting
pandas
xlsxwriter

# Utilities
python-dotenv
